Refactor media reducer.
