+++ packages/story-editor/src/app/media/local/reducers/addAudioProcessing.js
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+function addAudioProcessing(state, { id }) {
+  if (!id || state.audioProcessing.includes(id)) {
+    return state;
+  }
+  return {
+    ...state,
+    audioProcessing: [...state.audioProcessing, id],
+  };
+}
+
+export default addAudioProcessing;
