+++ packages/story-editor/src/app/media/local/reducers/index.js
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+export { default as setupState } from './setupState';
+export { default as fetchMedia } from './fetchMedia';
+export { default as setMediaType } from './setMediaType';
+export { default as setSearchTerm } from './setSearchTerm';
+export { default as resetFilters } from './resetFilters';
+export { default as setMedia } from './setMedia';
+export { default as prependMedia } from './prependMedia';
+export { default as addPosterProcessing } from './addPosterProcessing';
+export { default as removePosterProcessing } from './removePosterProcessing';
+export { default as addAudioProcessing } from './addAudioProcessing';
+export { default as removeAudioProcessing } from './removeAudioProcessing';
